+++ packages/media/src/index.ts
@@ -1,5 +1,6 @@
 export * from './orientation';
 export * from './validation';
+export * from './logo';
 export * from './probe';
 export * from './transcode';
 export * from './checksum';